renamed repo.rs to deps.rs
