+++ hatch/src/cli/commands/new.rs
@@ -1,7 +1,7 @@
 use std::fs;
 use clap::{ App, SubCommand, Arg, ArgMatches };
 use cli::commands::{ Command, parse_deps_from_cli };
-use repo::{ modules_path, hatchfile_path, clone_project_deps };
+use deps::{ modules_path, hatchfile_path, clone_project_deps };
 use project::{ Project, BuildConfig, ProjectKind, LibraryKind, Arch, Target, Dependency };
 use hatch_error::{ HatchResult, ResultExt };
 use task;
@@ -41,11 +41,11 @@ impl<'new> New {
     }
   }
 
-  fn construct_deps_string(&self, deps: &Vec<Dependency>) -> String {
-    if deps.is_empty() {
+  fn construct_deps_string(&self, project_deps: &Vec<Dependency>) -> String {
+    if project_deps.is_empty() {
       String::new()
     } else {
-      String::from("deps:\n") + deps.iter().map(|d| {
+      String::from("deps:\n") + project_deps.iter().map(|d| {
         format!("  {}: {}\n", d.name(), d.url())
       }).collect::<String>().as_str()
     }
@@ -139,15 +139,15 @@ impl<'command> Command<'command> for New {
       fs::create_dir_all(dir_path.join("test").join("src"))?;
       fs::create_dir(dir_path.join("test").join("target"))?;
 
-      let deps = deps_from_cli.into_iter().map(|repo| {
+      let project_deps = deps_from_cli.into_iter().map(|repo| {
         Dependency::new(repo)
       }).collect::<Vec<_>>();
 
-      if !deps.is_empty() {
-        clone_project_deps(modules_path(&dir_path).as_path(), &deps)?;
+      if !project_deps.is_empty() {
+        clone_project_deps(modules_path(&dir_path).as_path(), &project_deps)?;
       }
 
-      let includes = self.construct_deps_string(&deps);
+      let includes = self.construct_deps_string(&project_deps);
 
       let compiler: String = String::from("g++");
       let compiler_flags: Vec<String> = vec![String::from("-c")];
@@ -171,7 +171,7 @@ impl<'command> Command<'command> for New {
       let project = Project::new(name,
                                  version,
                                  config,
-                                 deps,
+                                 project_deps,
                                  dir_path.to_owned());
       task::generate_assets(&project)?;
