+++ hatch/src/lib.rs
@@ -16,7 +16,7 @@ pub mod hatch_error;
 pub mod yaml;
 pub mod assets;
 pub mod task;
-pub mod repo;
+pub mod deps;
 
 #[cfg(test)]
 mod tests {
